python/requirements.txt
Flask==0.12.1
pymongo==3.5.0
paho-mqtt==1.2.3
flasgger==0.9.2
Pillow==6.0.0
ImageHash==4.0.0